Ignoring patchsets that are Drafts
The bot catches events from all the patches and send updates to
IRC based on the configuration, however if a patch is in draft
status it is a work in progress and not meant to be seen until
it has been published.
This commit adds a condition to send a message to IRC regarding
a patch only if the patch is not in draft status. This will help
avoid spaming the IRC channel with draft patches.